Nachtrag Feature #174

+++ trunk/additional/sqldumps/tblTemporaryDoc.sql
@@ -8,7 +8,7 @@ CREATE TABLE ###TBLPREFIX###tblTemporaryDoc (
   DocumentID bigint(20) unsigned NOT NULL default '0',
   DocumentObject longtext NOT NULL,
   DocTable enum('tblFile','tblObjectFiles') NOT NULL default '',
-  UnixTimestamp bigint(20) unsigned NOT NULL default '0',
+  UnixTimestamp int(10) unsigned NOT NULL default '0',
   Active tinyint(1) unsigned NOT NULL default '0',
   PRIMARY KEY (`DocTable`,`DocumentID`,`Active`)
 ) ENGINE=MyISAM;
